Separate get info funcs; small updates
